better theme support

+++ bin/.profile
@@ -519,6 +519,7 @@ fi
 if [ "$PLATFORM" == "macosx" ]; then
   	alias xce='open -a /Applications/Xcode.app'
   	alias HH='cd ~/Desktop/Stuff/HH'
+  	alias themes='~/clanmills/files/themes'
   	if [ -e /opt/local/bin/svn ]; then
   		alias svn='/opt/local/bin/svn'
   	fi
@@ -544,10 +545,15 @@ if [ "$PLATFORM" == "linux" ]; then
   	alias svn='/usr/local/bin/svn'
 fi
 
-themes=/cygdrive/c/Users/rmills/AppData/Local/Microsoft/BingDesktop/themes/
-themez=/cygdrive/z/Users/rmills/clanmills/files/themes/
-if [ -e $themes ]; then
-	(cd $themes; for i in *.jpg;do if [ ! -e $themez/$i ]; then echo $i ; cp $i $themez ; fi; done)
+if [ "$PLATFORM" == "cygwin" ]; then
+	themes=/cygdrive/c/Users/rmills/AppData/Local/Microsoft/BingDesktop/themes/
+	themez=/cygdrive/z/Users/rmills/clanmills/files/themes/
+	if [ ! -e $themez ]; then
+		mkdir -p $themez 2>/dev/null >/dev/null
+	fi
+	if [[ -e $themes && -e $themez ]]; then
+		(cd $themes; for i in *.jpg;do if [ ! -e $themez/$i ]; then echo $i ; cp $i $themez ; fi; done)
+	fi
 fi
 
 cd ~/
